agent/gaia_benchmark/requirements.txt
requests
markdownify
youtube-transcript-api
pdfminer.six
mammoth
python-pptx
beautifulsoup4
pillow
pathvalidate
puremagic
serpapi